api/requirements.txt
# FastAPI
fastapi==0.65.2
starlette==0.14.2
uvicorn==0.14.0
pydantic[dotenv]
alembic==1.6.5
psycopg2-binary==2.9.1
fastapi-sqlalchemy==0.2.1
python-dotenv==0.18.0